Print effective machine name in verbose mode

+++ clfsm/examples/PingPongCLFSM.machine/State_Pong_OnEntry.mm
@@ -5,5 +5,5 @@
         
   stateName = "Pong";
   currentState = 1;
-        cout << stateName << " " << fmod((double)current_time_in_microseconds() / 1000000.0L, 100) << endl;
+    cout << stateName << " " << static_cast<int>(fmod(static_cast<double>(current_time_in_microseconds() / 1000000.0L), 100)) << endl;
 }
